requirements.txt
alembic==1.14.1
async-timeout==5.0.1
babel==2.17.0
bidict==0.23.1
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
Flask==3.1.0
flask-babel==4.0.0
Flask-Migrate==4.1.0
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
gevent==25.5.1
greenlet==3.2.2
gunicorn==23.0.0
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
paho-mqtt==1.6.1
pyproject_hooks==1.2.0
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2025.1
RPi.bme280==0.2.4
simple-websocket==1.1.0
smbus2==0.5.0
SQLAlchemy==2.0.38
typing_extensions==4.12.2
Werkzeug==3.1.3
wsproto==1.2.0
zope.event==5.0
zope.interface==7.2
